+++ packages/agent/src/utils/forest-schema/generator-actions.ts
@@ -63,6 +63,10 @@ export default class SchemaGeneratorActions {
       SchemaGeneratorActions.setFieldsDefaultValue(fields);
     }
 
+    if (schema.submitButtonLabel?.length > 50) {
+      throw new Error('Submit button label must have less than 50 characters');
+    }
+
     return {
       id: `${collection.name}-${actionIndex}-${slug}`,
       name,
@@ -73,6 +77,8 @@ export default class SchemaGeneratorActions {
       redirect: null, // frontend ignores this attribute
       download: Boolean(schema.generateFile),
       fields,
+      description: schema.description,
+      submitButtonLabel: schema.submitButtonLabel,
       hooks: {
         load: !schema.staticForm,
         // Always registering the change hook has no consequences, even if we don't use it.