+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,19 @@
+## example [1.10.1](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.10.0...example@1.10.1) (2024-09-25)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.46.0
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.30
+* **@forestadmin/datasource-mongo:** upgraded to 1.3.19
+* **@forestadmin/datasource-mongoose:** upgraded to 1.7.19
+* **@forestadmin/datasource-sequelize:** upgraded to 1.9.9
+* **@forestadmin/datasource-sql:** upgraded to 1.14.25
+* **@forestadmin/datasource-toolkit:** upgraded to 1.39.0
+
 # example [1.10.0](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.9.2...example@1.10.0) (2024-09-23)